fix(ref): broken hrefs on the skin commands ref page
Fix broken links to the neighboring files.
`.winset`, `.output`, and `.sound` commands are pathed under
`ref/skin/<command>.dm>`, which is different from the
`ref/skin/command/<command>.dm` pathing in the official BYOND reference.

There's a typo in the href of the .sound command page in the official
reference, with a dot missing before the command name. This is also
addressed in this fix.